requirements.txt
Flask
pandas
PyYAML
xmltodict
toml
Werkzeug
Jinja2
click
itsdangerous
MarkupSafe